requirements_runpod.txt
# RunPod Serverless Requirements for F5-TTS API
# PyTorch is pre-installed in the base image

# RunPod SDK
runpod>=1.6.0

# Audio Processing
soundfile>=0.12.1
librosa>=0.10.1
noisereduce>=3.0.0
scipy>=1.11.0
numpy>=1.24.0
pydub>=0.25.1

# ML/DL Framework
accelerate>=0.25.0
ema-pytorch>=0.3.0
safetensors>=0.4.0
transformers>=4.44.0

# F5-TTS Dependencies
cached_path>=1.5.0
hydra-core>=1.3.0
omegaconf>=2.3.0
datasets>=2.15.0
einops>=0.7.0
vocos>=0.1.0
jieba>=0.42.1
pypinyin>=0.50.0
x_transformers>=1.0.0
torchdiffeq>=0.2.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
huggingface_hub>=0.21.0

# Logging
wandb>=0.16.0